move the entity name portion of the entity editor component into its own component
