+++ src/duckling/entityeditor/entity-name.component.scss
